requirements.txt
pdfplumber
pandas
openpyxl
textwrap3
google-generativeai
streamlit
protobuf==4.25.3
